add name filter
